Included Scatterering Models

SDWBA Scatterer Models

SDWBA.Models.krill_mcgeehee — Constant.

Krill shape from McGeehee et al. 1998, 'Effects of orientation on acoustic scattering from Antarctic krill at 120 kHz.' Deep-Sea Research II 45(7), 1273-1294.

source

SDWBA.Models.krill_conti — Constant.

Krill shape from Conti and Demer, 2006, 'Improved parameterization of the SDWBA for estimating krill target strength.' ICES Journal of Marine Science 63(5), 928-935.

source

SDWBA.Models.sandeel — Constant.

Generic sand eel/sand lance shape (Ammodytes spp.). Shape is taken from illustration in Bigelow and Schroeder, 'Fishes of the Gulf of Maine.' Material properties from Yasuma et al. (2009), 'Density and sound-speed contrasts, and target strength of Japanese sandeel Ammodytes personatus.' Fisheries Science 75 (3) 545-552.

source

SDWBA.Models.calanoid_copepod — Constant.

Generic calanoid copepod shape, 1 mm long, with g and h for "typical" marine zooplankton (both 1.04, as per Stanton and Chu 2000, 'Review and recommendations for the modelling of acoustic scattering by fluid-like elongated zooplankton: euphausiids and copepods.' ICES Journal of Marine Science 57, 793-807).

source

SDWBA.Models.daphnia — Constant.

Generic Daphnia shape used in Warren et al. 2016, 'Measuring the distribution, abundance, and biovolume of zooplankton in an oligotrophic freshwater lake with a 710 kHz scientific echosounder.' Limnology and Oceanography: Methods

source

SDWBA.Models.daphnia2 — Constant.

Alternative (more realistic?) daphnia shape

source

SDWBA.Models.nauplius — Constant.

Generic nauplius larva shape, 0.5 mm long, with g and h for "typical" marine zooplankton (both 1.04, as per Stanton and Chu 2000, 'Review and recommendations for the modelling of acoustic scattering by fluid-like elongated zooplankton: euphausiids and copepods.' ICES Journal of Marine Science 57, 793-807).

source